+++ pkg/list/geo/dcg/util-fn/midpoint.ts
@@ -10,8 +10,8 @@ export const FN_MIDPOINT = new FnDist(
     "point32",
     (a, b) =>
       pt(
-        div(a.value.x.add(b.value.x), real(2)),
-        div(a.value.y.add(b.value.y), real(2)),
+        div(a.value.x.add(b.value.x), int(2)),
+        div(a.value.y.add(b.value.y), int(2)),
       ),
     (_, a, b) => `((${a.expr} + ${b.expr}) / 2.0)`,
     "midpoint((1,4),(3,8))=(2,6)",
@@ -20,7 +20,7 @@ export const FN_MIDPOINT = new FnDist(
     ["segment"],
     "point32",
     ({ value: [a, b] }) =>
-      pt(div(a.x.add(b.x), real(2)), div(a.y.add(b.y), real(2))),
+      pt(div(a.x.add(b.x), int(2)), div(a.y.add(b.y), int(2))),
     (ctx, ar) => {
       const a = ctx.cache(ar)
       return `((${a}.xy + ${a}.zw) / 2.0)`
